requirements_full.txt
requests
apscheduler
astral
pillow
pandas
fastapi
scikit-learn
scikit-image
smbus
pyserial
gpiozero
spidev